using the user.id taken from the token

+++ backend/controllers/goalController.js
@@ -1,6 +1,6 @@
 import asyncHandler from 'express-async-handler'
 import Goal from '../models/goalModel.js'
-
+import User from '../models/userModel.js'
 
 
 
@@ -12,7 +12,7 @@ import Goal from '../models/goalModel.js'
 // @route       Get /api/goals
 // @access      Private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find()
+    const goals = await Goal.find({ user: req.user.id })
 
     res.status(200).json(goals)
 })
@@ -35,7 +35,7 @@ const setGoals = asyncHandler(async (req, res) => {
 
     const goal = await Goal.create({
         text: req.body.text,
-        // user: req.body.id
+        user: req.user.id           // we get id from the protect function
     })
 
     res.status(200).json(goal)
@@ -56,6 +56,21 @@ const updateGoal = asyncHandler(async (req, res) => {
     if (!goal) {
         res.status(400)
         throw new Error('Goal not found')
+    }
+
+    // we will get user.id from the token
+    const user = await User.findById(req.user.id)
+
+    // check for user
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // make sure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorize')
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -82,6 +97,22 @@ const deleteGoals = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error('Goal not found')
     }
+
+    // we will get user.id from the token
+    const user = await User.findById(req.user.id)
+
+    // check for user
+    if (!user) {
+        res.status(401)
+        throw new Error('User not found')
+    }
+
+    // make sure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error('User not authorize')
+    }
+
 
     await goal.deleteOne()
